Change transformations to implement Cursor
